+++ src/pixel/include/hictk/pixel.hpp
@@ -20,9 +20,9 @@ class Chromosome;
 
 template <typename N>
 struct ThinPixel {
-  std::size_t bin1_id{std::numeric_limits<std::size_t>::max()};  // NOLINT
-  std::size_t bin2_id{std::numeric_limits<std::size_t>::max()};  // NOLINT
-  N count{};                                                     // NOLINT
+  std::uint64_t bin1_id{std::numeric_limits<std::uint64_t>::max()};  // NOLINT
+  std::uint64_t bin2_id{std::numeric_limits<std::uint64_t>::max()};  // NOLINT
+  N count{};                                                         // NOLINT
 
   [[nodiscard]] explicit operator bool() const noexcept;
   [[nodiscard]] bool operator==(const ThinPixel &other) const noexcept;
